BatteryLights: Fix disabled lights when fully charged default
According to def_battery_light_full_charge_disabled in
/packages/apps/LineageParts/res/values/bools.xml the setting should be
false by default, but currently this applies only to the switch in the
preferences. As long as the setting is not defined, its default is
defined in the call to LineageSettings.System.getIntForUser.

This commit aligns the default in the latter with bools.xml.

Fixes: https://gitlab.com/LineageOS/issues/android/-/issues/8486
Change-Id: If94b2f50bbda7cbd42af2b587e75be9a56b8876e